fix(actions): refactored IaC plugin
refactored IaC plugin
