+++ packages/backend/src/index.ts
@@ -5,7 +5,7 @@ import { createBackendModule } from '@backstage/backend-plugin-api';
 import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
 import { createMicroserviceAction } from './plugins/sourceloop-ms';
 import { createScaffoldAction } from './plugins/sourceloop-scaffold';
-import { modifyIaCModules } from '../../../plugins/scaffolder-backend-module-prepare-iac-modules';
+import { modifyIaCModules } from './plugins/iac-scaffold';
 
 const scaffolderModuleCustomExtensions = createBackendModule({
   pluginId: 'scaffolder', // name of the plugin that the module is targeting
@@ -19,8 +19,8 @@ const scaffolderModuleCustomExtensions = createBackendModule({
         scaffolder.addActions(
           createMicroserviceAction(),
           createScaffoldAction(),
+          modifyIaCModules()
         );
-        scaffolder.addActions(modifyIaCModules());
       },
     });
   },
@@ -69,5 +69,4 @@ backend.add(import('@backstage-community/plugin-jenkins-backend'));
 backend.add(
   import('@internal/backstage-plugin-jenkins-with-reporting-backend-backend'),
 );
-backend.add(import('../../../plugins/scaffolder-backend-module-prepare-iac-modules/src/actions/prepare-iac'));
 backend.start();
