--- plugins/scaffolder-backend-module-prepare-iac-modules/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@internal/backstage-plugin-scaffolder-backend-module-prepare-iac-modules",
-  "description": "The prepare-iac-modules module for @backstage/plugin-scaffolder-backend",
-  "version": "0.1.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "publishConfig": {
-    "access": "public",
-    "main": "dist/index.cjs.js",
-    "types": "dist/index.d.ts"
-  },
-  "backstage": {
-    "role": "backend-plugin-module"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean",
-    "prepack": "backstage-cli package prepack",
-    "postpack": "backstage-cli package postpack"
-  },
-  "dependencies": {
-    "@backstage/plugin-scaffolder-node": "^0.4.11"
-  },
-  "devDependencies": {
-    "@backstage/cli": "^0.27.1"
-  },
-  "files": [
-    "dist"
-  ]
-}
