--- plugins/scaffolder-backend-module-prepare-iac-modules/src/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './prepare-iac';
