--- plugins/scaffolder-backend-module-prepare-iac-modules/src/actions/prepare-iac/index.ts
@@ -1,7 +1,0 @@
-import { scaffolderModule } from './module';
-
-/*
-  @deprecated - this way of importing modules will soon be unsupported, and you should use `backend.add(import(...))` instead.
-*/
-export { modifyIaCModules } from './iac';
-export default scaffolderModule;
