--- plugins/scaffolder-backend-module-prepare-iac-modules/src/actions/prepare-iac/module.ts
@@ -1,21 +1,0 @@
-import { createBackendModule } from "@backstage/backend-plugin-api";
-import { scaffolderActionsExtensionPoint  } from '@backstage/plugin-scaffolder-node/alpha';
-import { modifyIaCModules } from "./iac";
-
-/**
- * A backend module that registers the action into the scaffolder
- */
-export const scaffolderModule = createBackendModule({
-  moduleId: 'acme:example',
-  pluginId: 'scaffolder-iac',
-  register({ registerInit }) {
-    registerInit({
-      deps: {
-        scaffolderActions: scaffolderActionsExtensionPoint
-      },
-      async init({ scaffolderActions}) {
-        scaffolderActions.addActions(modifyIaCModules());
-      }
-    });
-  },
-})
