--- plugins/scaffolder-backend-module-prepare-iac-modules/src/index.ts
@@ -1,8 +1,0 @@
-/***/
-/**
- * The prepare-iac-modules module for @backstage/plugin-scaffolder-backend.
- *
- * @packageDocumentation
- */
-
-export * from './actions';
